style(game-header): move tier verdict above the dial to stop it clipping
PLATINUM overflowed the 118px confidence dial ('LATINUM'). Put the full tier
name as a label ABOVE the dial (readable at any tier, colored) and let the dial
center show just the big confidence % + caption. Drops the short/full media-query
hack.
