+++ css/app/game-header.css
@@ -279,26 +279,18 @@
   min-width: 0;
 }
 .grp-row { display: flex; align-items: center; gap: 18px; flex: 1 1 auto; }
+/* The tier verdict is a full-name label ABOVE the dial (readable at any tier,
+   no clipping inside the 118px circle); the dial itself shows only the
+   confidence % + caption. */
+.grp-dial-block { flex: 0 0 auto; display: flex; flex-direction: column; align-items: center; gap: 6px; }
+.grp-dial-verdict { font-size: 0.82rem; font-weight: 800; letter-spacing: 0.07em; text-transform: uppercase; line-height: 1; text-align: center; }
 .grp-dial { position: relative; flex: 0 0 auto; width: 118px; height: 118px; }
 .grp-dial svg { width: 100%; height: 100%; display: block; }
 .grp-dial-ctr {
   position: absolute; inset: 0;
   display: flex; flex-direction: column; align-items: center; justify-content: center; gap: 1px;
 }
-/* Tier is the headline -- what the game IS. Short form (PLAT/GOLD/
-   SILV/BRON/BORK) is the SAFE DEFAULT so even a stale-cache browser
-   doesn't overflow the 118px dial with PLATINUM. Desktop >=1024px
-   opts back into the full form. Ordering the short span first also
-   means it's what any missing-media-query fallback shows. */
-.grp-dial-tier { font-size: 1.35rem; font-weight: 800; letter-spacing: 0.06em; text-transform: uppercase; line-height: 1; }
-.grp-dial-tier-short { display: inline; }
-.grp-dial-tier-full  { display: none; }
-@media (min-width: 1024px) {
-  .grp-dial-tier-short { display: none; }
-  .grp-dial-tier-full  { display: inline; }
-  .grp-dial-tier       { font-size: 1.1rem; letter-spacing: 0.04em; }
-}
-.grp-dial-pct { font-size: 0.9rem; font-weight: 700; color: var(--strong); font-variant-numeric: tabular-nums; line-height: 1.1; margin-top: 4px; }
+.grp-dial-pct { font-size: 1.5rem; font-weight: 800; color: var(--strong); font-variant-numeric: tabular-nums; line-height: 1; }
 .grp-dial-cap { font-size: 0.5rem; letter-spacing: 0.06em; text-transform: uppercase; color: var(--muted); font-family: ui-monospace, "SFMono-Regular", Menlo, Consolas, monospace; }
 .grp-bars { flex: 1 1 0; min-width: 0; display: flex; flex-direction: column; gap: 6px; }
 .grp-bar {
